chore(api): align clause types (#1416)
