+++ api/transactions_types.go
@@ -6,8 +6,6 @@
 package api
 
 import (
-	"fmt"
-
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/common/math"
 
@@ -15,25 +13,6 @@ import (
 	"github.com/vechain/thor/v2/thor"
 	"github.com/vechain/thor/v2/tx"
 )
-
-// ConvertClause convert a raw clause into a json format clause
-func ConvertClause(c *tx.Clause) Clause {
-	return Clause{
-		To:    c.To(),
-		Value: (*math.HexOrDecimal256)(c.Value()),
-		Data:  hexutil.Encode(c.Data()),
-	}
-}
-
-func (c *Clause) String() string {
-	return fmt.Sprintf(`Clause(
-		To    %v
-		Value %v
-		Data  %v
-		)`, c.To,
-		c.Value,
-		c.Data)
-}
 
 type RawTx struct {
 	Raw string `json:"raw"`
@@ -88,20 +67,6 @@ type Output struct {
 	ContractAddress *thor.Address `json:"contractAddress"`
 	Events          []*Event      `json:"events"`
 	Transfers       []*Transfer   `json:"transfers"`
-}
-
-// Event event.
-type Event struct {
-	Address thor.Address   `json:"address"`
-	Topics  []thor.Bytes32 `json:"topics"`
-	Data    string         `json:"data"`
-}
-
-// Transfer transfer log.
-type Transfer struct {
-	Sender    thor.Address          `json:"sender"`
-	Recipient thor.Address          `json:"recipient"`
-	Amount    *math.HexOrDecimal256 `json:"amount"`
 }
 
 // ConvertReceipt convert a raw clause into a jason format clause
